Rename AspNetCoreUtils project and namespace to ApiCore
