+++ shared-kernel/ApiCore/ApiCore.csproj
@@ -4,8 +4,8 @@
         <TargetFramework>net7.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <AssemblyName>PlatformPlatform.SharedKernel.AspNetCoreUtils</AssemblyName>
-        <RootNamespace>PlatformPlatform.SharedKernel.AspNetCoreUtils</RootNamespace>
+        <AssemblyName>PlatformPlatform.SharedKernel.ApiCore</AssemblyName>
+        <RootNamespace>PlatformPlatform.SharedKernel.ApiCore</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
