+++ shared-kernel/ApiCore/Middleware/GlobalExceptionHandlerMiddleware.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 
-namespace PlatformPlatform.SharedKernel.AspNetCoreUtils.Middleware;
+namespace PlatformPlatform.SharedKernel.ApiCore.Middleware;
 
 public sealed class GlobalExceptionHandlerMiddleware : IMiddleware
 {
